# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type BatchMetadataStorageCondition =
  (typeof BatchMetadataStorageCondition)[keyof typeof BatchMetadataStorageCondition];

export const BatchMetadataStorageCondition = {
  Hot: "Hot",
  Cold: "Cold",
  Room_Temperature: "Room Temperature",
} as const;

export interface BatchMetadata {
  name: string;
  expiryDate: string;
  quantity: number;
  composition: string;
  storageCondition: BatchMetadataStorageCondition;
  manufacturerAddress: string;
  batchId: string;
  uploadedAt?: string;
  /** @nullable */
  cid?: string | null;
}

export type UploadMetadataBodyStorageCondition =
  (typeof UploadMetadataBodyStorageCondition)[keyof typeof UploadMetadataBodyStorageCondition];

export const UploadMetadataBodyStorageCondition = {
  Hot: "Hot",
  Cold: "Cold",
  Room_Temperature: "Room Temperature",
} as const;

export interface UploadMetadataBody {
  name: string;
  expiryDate: string;
  quantity: number;
  composition: string;
  storageCondition: UploadMetadataBodyStorageCondition;
  manufacturerAddress: string;
  batchId: string;
}

export interface UploadMetadataResponse {
  cid: string;
  ipfsUrl: string;
  pinataUrl: string;
}

export type SupplyChainEventEventType =
  (typeof SupplyChainEventEventType)[keyof typeof SupplyChainEventEventType];

export const SupplyChainEventEventType = {
  Registered: "Registered",
  Transferred: "Transferred",
  Accepted: "Accepted",
  Sold: "Sold",
} as const;

export interface SupplyChainEvent {
  eventType: SupplyChainEventEventType;
  /** @nullable */
  fromAddress?: string | null;
  toAddress: string;
  timestamp: number;
  batchId: string;
  /** @nullable */
  transactionHash?: string | null;
  roleName: string;
}

export type VerificationResultCurrentStatus =
  (typeof VerificationResultCurrentStatus)[keyof typeof VerificationResultCurrentStatus];

export const VerificationResultCurrentStatus = {
  Active: "Active",
  Sold: "Sold",
  Recalled: "Recalled",
} as const;

export interface VerificationResult {
  batchId: string;
  metadata: BatchMetadata;
  events: SupplyChainEvent[];
  currentStatus: VerificationResultCurrentStatus;
  currentHolder: string;
  currentHolderRole: string;
  isAuthentic: boolean;
}

export type QrGenerateBodyType =
  (typeof QrGenerateBodyType)[keyof typeof QrGenerateBodyType];

export const QrGenerateBodyType = {
  identity: "identity",
  batch: "batch",
} as const;

export interface QrGenerateBody {
  type: QrGenerateBodyType;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  batchId?: string | null;
  /** @nullable */
  label?: string | null;
}

export interface QrCodeResponse {
  qrCodeDataUrl: string;
  data: string;
  /** @nullable */
  label?: string | null;
}

export interface SystemStats {
  totalManufacturers: number;
  totalBatches: number;
  totalTransfers: number;
  totalSold: number;
  totalDistributors: number;
  totalPharmacies: number;
}