# lib/api-zod/src/generated/types/batchMetadata.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BatchMetadataStorageCondition } from "./batchMetadataStorageCondition";

export interface BatchMetadata {
  name: string;
  expiryDate: string;
  quantity: number;
  composition: string;
  storageCondition: BatchMetadataStorageCondition;
  manufacturerAddress: string;
  batchId: string;
  uploadedAt?: string;
  /** @nullable */
  cid?: string | null;
}