# lib/api-zod/src/generated/types/qrCodeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QrCodeResponse {
  qrCodeDataUrl: string;
  data: string;
  /** @nullable */
  label?: string | null;
}