# lib/api-zod/src/generated/types/qrGenerateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QrGenerateBodyType } from "./qrGenerateBodyType";

export interface QrGenerateBody {
  type: QrGenerateBodyType;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  batchId?: string | null;
  /** @nullable */
  label?: string | null;
}