# lib/api-zod/src/generated/types/supplyChainEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SupplyChainEventEventType } from "./supplyChainEventEventType";

export interface SupplyChainEvent {
  eventType: SupplyChainEventEventType;
  /** @nullable */
  fromAddress?: string | null;
  toAddress: string;
  timestamp: number;
  batchId: string;
  /** @nullable */
  transactionHash?: string | null;
  roleName: string;
}