# lib/api-zod/src/generated/types/systemStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SystemStats {
  totalManufacturers: number;
  totalBatches: number;
  totalTransfers: number;
  totalSold: number;
  totalDistributors: number;
  totalPharmacies: number;
}