# lib/api-zod/src/generated/types/uploadMetadataBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UploadMetadataBodyStorageCondition } from "./uploadMetadataBodyStorageCondition";

export interface UploadMetadataBody {
  name: string;
  expiryDate: string;
  quantity: number;
  composition: string;
  storageCondition: UploadMetadataBodyStorageCondition;
  manufacturerAddress: string;
  batchId: string;
}