# lib/api-zod/src/generated/types/uploadMetadataResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadMetadataResponse {
  cid: string;
  ipfsUrl: string;
  pinataUrl: string;
}