# lib/api-zod/src/generated/types/verificationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PharmaChain API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BatchMetadata } from "./batchMetadata";
import type { SupplyChainEvent } from "./supplyChainEvent";
import type { VerificationResultCurrentStatus } from "./verificationResultCurrentStatus";

export interface VerificationResult {
  batchId: string;
  metadata: BatchMetadata;
  events: SupplyChainEvent[];
  currentStatus: VerificationResultCurrentStatus;
  currentHolder: string;
  currentHolderRole: string;
  isAuthentic: boolean;
}